% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccountDistribution.R
\name{dataAccountDistribution}
\alias{dataAccountDistribution}
\title{Get data about a set of countries used as support for distribution graphs}
\usage{
dataAccountDistribution(year = 2019, exclude_country = "")
}
\arguments{
\item{year}{number year to compare countries}

\item{exclude_country}{string indicating the country name that will be excluded from the result dataset (country name in PT)}
}
\value{
data about a set of countries.
}
\description{
Get data about a set of countries used as support for distribution graphs
}
\examples{
dataAccountDistribution(year=2019)
}
