% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccountEvolutionByCountries.R
\name{dataAccountEvolutionByCountries}
\alias{dataAccountEvolutionByCountries}
\title{Get data about the evolution of an account values for selected countries}
\usage{
dataAccountEvolutionByCountries(
  account = "Despesa",
  continent = "",
  group = "",
  level = "",
  country = "",
  zone = "",
  strategy = ""
)
}
\arguments{
\item{account}{string with account name}

\item{continent}{string with continent name}

\item{group}{string with economic group}

\item{level}{string with development level}

\item{country}{string with country name in PT}

\item{zone}{string with economic zone}

\item{strategy}{string with strategic issue}
}
\value{
data about the evolution of an account values for selected countries.
}
\description{
Get data about the evolution of an account values for selected countries
}
\examples{
dataAccountEvolutionByCountries(account="Despesa", group = "G20")
}
