% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCompleteTimeSeries.R
\name{dataCompleteTimeSeries}
\alias{dataCompleteTimeSeries}
\title{Get data about account values for a set of countries used as support for time series graphs}
\usage{
dataCompleteTimeSeries(account = "Despesa", exclude_country = "")
}
\arguments{
\item{account}{string with account name}

\item{exclude_country}{string indicating the country name that will be excluded from the result dataset (country name in PT)}
}
\value{
data about account values for a set of countries.
}
\description{
Get data about account values for a set of countries used as support for time series graphs
}
\examples{
dataCompleteTimeSeries(account="Despesa")
}
