% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHeatMap.R
\name{dataHeatMap}
\alias{dataHeatMap}
\title{Get data about account values for a set of countries used as support for heat map graphs}
\usage{
dataHeatMap(account = "Despesa", exclude_country = "")
}
\arguments{
\item{account}{string with account name}

\item{exclude_country}{string indicating the country name that will be excluded from the result dataset (country name in PT)}
}
\value{
data about account values for a set of countries.
}
\description{
Get data about account values for a set of countries used as support for heat map graphs
}
\examples{
dataHeatMap(account="Despesa")
}
