% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountries.R
\name{getCountries}
\alias{getCountries}
\title{Get all the available countries}
\usage{
getCountries()
}
\value{
available countries.
}
\description{
Get all the available countries
}
\examples{
getCountries()
}
