% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDevelopmentLevels.R
\name{getDevelopmentLevels}
\alias{getDevelopmentLevels}
\title{Get the development levels used in GFS}
\usage{
getDevelopmentLevels()
}
\value{
development levels.
}
\description{
Get the development levels used in GFS
}
\examples{
getDevelopmentLevels()
}
