% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEconomicGroups.R
\name{getEconomicGroups}
\alias{getEconomicGroups}
\title{Get all the economic groups used in GFS}
\usage{
getEconomicGroups()
}
\value{
economic groups.
}
\description{
Get all the economic groups used in GFS
}
\examples{
getEconomicGroups()
}
