% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEconomicZones.R
\name{getEconomicZones}
\alias{getEconomicZones}
\title{Get all the economic zones used in GFS}
\usage{
getEconomicZones()
}
\value{
economic zones.
}
\description{
Get all the economic zones used in GFS
}
\examples{
getEconomicZones()
}
