% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrategicIssues.R
\name{getStrategicIssues}
\alias{getStrategicIssues}
\title{Get all the strategic issues used in GFS}
\usage{
getStrategicIssues()
}
\value{
strategic issues.
}
\description{
Get all the strategic issues used in GFS
}
\examples{
getStrategicIssues()
}
