% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphAccountDistribution.R
\name{graphAccountDistribution}
\alias{graphAccountDistribution}
\title{Generate a graph showing a the distribution of account values for a set of countries}
\usage{
graphAccountDistribution(.data, selected_country = "Brasil", text_max = FALSE)
}
\arguments{
\item{.data}{tibble data about the distribution of account values for a set of countries}
}
\value{
box plot graph.
}
\description{
Generate a graph showing a the distribution of account values for a set of countries
}
\examples{
dataAccountDistribution(year=2019) \%>\% graphAccountDistribution(selected_country="Brasil", text_max = FALSE )
}
