% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphAccountEvolutionByCountries.R
\name{graphAccountEvolutionByCountries}
\alias{graphAccountEvolutionByCountries}
\title{Generate a graph showing the evolution of account values for selected countries}
\usage{
graphAccountEvolutionByCountries(.data)
}
\arguments{
\item{.data}{tibble data about the evolution of account values for selected countries}
}
\value{
dynamic graph the evolution of account values for selected countries.
}
\description{
Generate a graph showing the evolution of account values for selected countries
}
\examples{
dataAccountEvolutionByCountries(account="Despesa", group = "G20") \%>\% graphAccountEvolutionByCountries()
}
