% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphCompleteTimeSeries.R
\name{graphCompleteTimeSeries}
\alias{graphCompleteTimeSeries}
\title{Generate a graph showing a time series about the evolution of account values for selected countries}
\usage{
graphCompleteTimeSeries(.data, selected_country = "Brasil", text_max = FALSE)
}
\arguments{
\item{.data}{tibble data about the evolution of account values for selected countries}
}
\value{
time series graph.
}
\description{
Generate a graph showing a time series about the evolution of account values for selected countries
}
\examples{
dataCompleteTimeSeries(account="Despesa") \%>\% graphCompleteTimeSeries(selected_country="Brasil", text_max = FALSE )
}
